requirements.txt
redis
rq
slack-bolt
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
gitpython
python-dotenv
requests
pytest
openai
crewai
firecrawl-py
langchain
langchain-google-genai
langchain-openai
numpy